# compact show for operators