--- a/src/show.jl
+++ b/src/show.jl
@@ -46,9 +46,11 @@ function show(io::IO, ::MIME"text/plain", @nospecialize(B::Operator); header::Bo
 
     iocompact = haskey(io, :compact) ? io : IOContext(io, :compact=>true)
 
+    sz1_B, sz2_B = size(B)
+
     if !isambiguous(domainspace(B)) && (eltype(B) <: Number)
-        println()
-        if isbanded(B) && isinf(size(B,1)) && isinf(size(B,2))
+        println(io)
+        if isbanded(B) && isinf(sz1_B) && isinf(sz2_B)
             BM=B[1:10,1:10]
 
             M=Matrix{Union{eltype(B), PrintShow}}(undef,11,11)
@@ -65,48 +67,55 @@ function show(io::IO, ::MIME"text/plain", @nospecialize(B::Operator); header::Bo
             end
 
             print_array(iocompact, M)
-        elseif isinf(size(B,1)) && isinf(size(B,2))
+        elseif isinf(sz1_B) && isinf(sz2_B)
             BM=B[1:10,1:10]
 
             M=Matrix{Union{eltype(B), PrintShow}}(undef,11,11)
-            for k=1:10,j=1:10
-                M[k,j]=BM[k,j]
+            for I in CartesianIndices((1:10, 1:10))
+                M[I]=BM[Tuple(I)...] # not certain if indexing with CartesianIndex is implemented
             end
 
             M[1,end]=PrintShow('⋯')
             M[end,1]=PrintShow('⋮')
 
             for k=2:11
-                M[k,end]=M[end,k]=PrintShow('⋱')
+                M[k,end]=PrintShow('⋱')
+            end
+            for k=2:11
+                M[end,k]=PrintShow('⋱')
             end
 
             print_array(iocompact, M)
-        elseif isinf(size(B,1))
-            BM=B[1:10,1:size(B,2)]
+        elseif isinf(sz1_B)
+            sz2int = Int(sz2_B)::Int
+            BM=B[1:10,1:sz2int]
 
-            M=Matrix{Union{eltype(B), PrintShow}}(undef,11,size(B,2))
-            for k=1:10,j=1:size(B,2)
-                M[k,j]=BM[k,j]
+            M=Matrix{Union{eltype(B), PrintShow}}(undef,11,sz2int)
+            for I in CartesianIndices((1:10, 1:10))
+                M[I]=BM[Tuple(I)...]
             end
-            for k=1:size(B,2)
+            for k=1:sz2int
                 M[end,k]=PrintShow('⋮')
             end
 
             print_array(iocompact, M)
-        elseif isinf(size(B,2))
-            BM=B[1:size(B,1),1:10]
+        elseif isinf(sz2_B)
+            sz1int = Int(sz1_B)::Int
+            BM=B[1:sz1int,1:10]
 
-            M=Matrix{Union{eltype(B), PrintShow}}(undef,size(B,1),11)
-            for k=1:size(B,1),j=1:10
-                M[k,j]=BM[k,j]
+            M=Matrix{Union{eltype(B), PrintShow}}(undef,sz1int,11)
+            for I in CartesianIndices((1:10, 1:10))
+                M[I]=BM[Tuple(I)...]
             end
-            for k=1:size(B,1)
+            for k=1:sz1int
                 M[k,end]=PrintShow('⋯')
             end
 
             print_array(iocompact, M)
         else
-            print_array(iocompact, AbstractMatrix(B)[1:size(B,1),1:size(B,2)])
+            sz1int = Int(sz1_B)::Int
+            sz2int = Int(sz2_B)::Int
+            print_array(iocompact, AbstractMatrix(B)[1:sz1int,1:sz2int])
         end
     end
 end