--- a/test/SpacesTest.jl
+++ b/test/SpacesTest.jl
@@ -340,5 +340,12 @@ using StaticArrays
             @test Derivative(Chebyshev()) != Derivative(Chebyshev(), 2)
             @test Derivative(Chebyshev()) != Derivative(Legendre())
         end
+
+        @testset "SubOperator" begin
+            D = Derivative(Chebyshev())
+            S = @view D[1:10, 1:10]
+            @test rowrange(S, 1) == 2:2
+            @test colrange(S, 2) == 1:1
+        end
     end
 end