--- a/test/show.jl
+++ b/test/show.jl
@@ -1,4 +1,5 @@
 @testset "show" begin
+	io = IOBuffer()
 	@testset "Domain" begin
 		@testset "Segment" begin
 			s = Segment(0, 1)
@@ -33,7 +34,6 @@
 			D = Derivative()
 			dsum = ApproxFunBase.summarystr(D)
 			@test repr(D) == dsum
-			io = IOBuffer()
 			show(io, MIME"text/plain"(), D)
 			@test contains(String(take!(io)), dsum)
 
@@ -52,7 +52,6 @@
 		@testset "QuotientSpace" begin
 			Q = QuotientSpace(Dirichlet(ConstantSpace(0..1)))
 			@test startswith(repr(Q), "ConstantSpace(0..1) /")
-			io = IOBuffer()
 			show(io, MIME"text/plain"(), Q)
 			s = String(take!(io))
 			@test startswith(s, "ConstantSpace(0..1) /")