This makes the display of operators similar to that for matrices.
After this
```julia
julia> Derivative(NormalizedChebyshev())
DerivativeWrapper : NormalizedChebyshev() → Ultraspherical(1)
 ⋅  0.797885   ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       ⋅
 ⋅   ⋅        1.59577   ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       ⋅
 ⋅   ⋅         ⋅       2.39365   ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       ⋅
 ⋅   ⋅         ⋅        ⋅       3.19154   ⋅        ⋅        ⋅        ⋅        ⋅       ⋅
 ⋅   ⋅         ⋅        ⋅        ⋅       3.98942   ⋅        ⋅        ⋅        ⋅       ⋅
 ⋅   ⋅         ⋅        ⋅        ⋅        ⋅       4.78731   ⋅        ⋅        ⋅       ⋅
 ⋅   ⋅         ⋅        ⋅        ⋅        ⋅        ⋅       5.58519   ⋅        ⋅       ⋅
 ⋅   ⋅         ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       6.38308   ⋅       ⋅
 ⋅   ⋅         ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       7.18096  ⋅
 ⋅   ⋅         ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       ⋱
 ⋅   ⋅         ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅        ⋅       ⋱
```
This also makes the display times slightly faster by avoiding specializing on the types of the operators.
On master
```julia
julia> D = Derivative(NormalizedChebyshev());

julia> io = IOBuffer();

julia> @time show(io, MIME("text/plain"), D);

  4.935376 seconds (11.19 M allocations: 564.487 MiB, 5.19% gc time, 99.89% compilation time)

julia> D = Derivative(Chebyshev());

julia> @time show(io, MIME("text/plain"), D);

  0.123173 seconds (500.34 k allocations: 24.599 MiB, 99.54% compilation time)
```
This PR
```julia
julia> using ApproxFun

julia> io = IOBuffer();

julia> D = Derivative(NormalizedChebyshev());

julia> @time show(io, MIME("text/plain"), D);
  4.534671 seconds (10.72 M allocations: 533.026 MiB, 4.54% gc time, 99.81% compilation time)

julia> D = Derivative(Chebyshev());

julia> @time show(io, MIME("text/plain"), D);
  0.029711 seconds (37.16 k allocations: 2.003 MiB, 87.81% compilation time)
```